Added opportunity of rendering of multiple meshs from fbx file format and Added Texture manager

+++ src/render/directx.h
@@ -43,7 +43,7 @@ struct Direct_Write {
 inline int Direct_Write::get_text_width(const char *text)
 {
 	D2D1_SIZE_F size = get_text_size_in_pixels(text);
-	return size.width;
+	return static_cast<int>(size.width);
 }
 
 extern Direct_Write direct_write;
